refactor: Standardize votingEndTimestamp to uint48
Updates the `votingEndTimestamp` type to `uint48` across relevant contracts, including voting strategies, validators, Azorius, and mocks.

This change provides a significantly larger range for timestamps compared to `uint32` (preventing issues around the year 2106) while optimizing gas usage compared to `uint256`. The `currentTimestamp` in the `VotingPeriodEnded` event is also updated to `uint48` for consistency.

+++ contracts/deployables/account-abstraction/validators/LinearERC721VotingV1ValidatorV1.sol
@@ -23,7 +23,7 @@ interface ILinearERC721VotingV1 {
 
     function votingEndTimestamp(
         uint32 proposalId
-    ) external view returns (uint32);
+    ) external view returns (uint48);
 
     function getTokenWeight(
         address tokenAddress
